feat(DirectoryPage): refactor organization directory types and improve routing consistency
